+++ CHANGELOG.md
@@ -11,6 +11,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Add support for simulating errors.
 
+### Changed
+
+- Children passed to non-rendered components during shallow-rendering are now
+  present in the output, for consistency with how shallow rendering works in
+  React.
+
 ### Fixed
 
 - Shallow rendering now only renders the root element passed to `shallow`,